Refactor baselayout
