+++ src/app/App.tsx
@@ -1,6 +1,6 @@
 import { useEffect, FC } from 'react';
 import { migrateGoalProjectLinks } from '../entities/goal/model/migrateGoals';
-import { BasicLayout } from "@/shared/ui/BasicLayout"
+import { MainLayout } from "@/widgets/MainLayout"
 import { SettingsProvider } from "@/features/settings"
 import type { Platform } from './lib/platform';
 import { ThemeProvider } from '@/shared/theme/ThemeContext';
@@ -18,7 +18,7 @@ export const App: FC<AppProps> = ({ platform }) => {
     return (
         <ThemeProvider>
             <SettingsProvider>
-                <BasicLayout />
+                <MainLayout />
             </SettingsProvider>
         </ThemeProvider>
     )
