+++ src/shared/ui/BaseLayout/index.ts
@@ -1,0 +1,1 @@
+export { BaseLayout } from './BaseLayout';
