--- src/shared/ui/BasicLayout.css
@@ -1,193 +1,0 @@
-.basicLayout {
-    padding: 2rem;
-    /* max-width: 1200px; */
-    height: 100%;
-    margin: 0 auto;
-    overflow-x:scroll;
-    overflow-y: hidden;
-    z-index: 0;
-}
-.basicLayout__content, .basicLayout__content__taskList {
-    min-height: 90%;
-}
-@media (max-width: 768px) {
-    .basicLayout {
-        padding: 0;
-    }
-
-    .sidebar-collapsed {
-        min-width: 0px !important;
-        width: 0px !important;
-    }
-
-    .sidebar  {
-        min-width: 0px !important;
-    }
-    .basicLayout__content__lists {
-        margin-left: 0px !important;
-    }
-}
-
-.basicLayout__header {
-    margin-bottom: 2rem;
-    text-align: center;
-}
-
-.basicLayout__content {
-    display: flex;
-    flex-direction: row;
-    gap: 20px;
-    overflow: hidden;
-    position: relative;
-    transition: all 0.3s ease;
-}
-
-.sidebar {
-    position: relative;
-    min-width: 520px;
-    transition: all 0.3s ease;
-    background: var(--color-background);
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    margin-right: 20px;
-}
-
-.sidebar-collapsed {
-    min-width: 40px;
-    width: 40px;
-}
-
-.sidebar-toggle {
-    position: absolute;
-    right: -20px;
-    top: 20px;
-    width: 40px;
-    height: 40px;
-    background: var(--color-background);
-    border: 1px solid var(--color-background-alt);
-    border-radius: 50%;
-    cursor: pointer;
-    z-index: 10;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 20px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    transition: all 0.2s ease;
-    color: var(--color-text);
-}
-
-.sidebar-toggle:hover {
-    background: var(--color-background-alt);
-    transform: scale(1.1);
-}
-
-.sidebar-closed .basicLayout__content__lists {
-    margin-left: 40px;
-}
-
-.basicLayout__content__lists {
-    overflow-x: auto;
-    overflow-y: hidden;
-    cursor: grab;
-    user-select: none;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    scroll-behavior: smooth;
-    -webkit-overflow-scrolling: touch;
-    scrollbar-width: thin;
-    padding: 1rem 0;
-    flex: 1;
-    transition: margin-left 0.3s ease;
-}
-
-.basicLayout__content__lists:active {
-    cursor: grabbing;
-}
-
-.basicLayout__content__taskList {
-    display: flex;
-    flex-direction: row;
-    gap: 20px;
-    min-width: max-content;
-    padding: 0 1rem;
-}
-
-.basicLayout__content__projects {
-    width: 400px;
-    min-width: 300px;
-    background: var(--color-background);
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.basicLayout__content__goals {
-    width: 400px;
-    min-width: 300px;
-    background: var(--color-background);
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.basicLayout__content__rewards {
-    width: 400px;
-    min-width: 300px;
-    background: var(--color-background);
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.basicLayout__content__sheet {
-    width: 600px;
-    background: var(--color-background);
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    padding: 16px;
-}
-
-.basicLayout__content__brainDump, .brainDump {
-    width: fit-content;
-    margin-right: 10px;
-    height: fit-content;
-}
-
-.basicLayout__settings-button {
-    padding: 0.5rem 1rem;
-    background-color: var(--color-background-alt);
-    color: var(--color-text);
-    border: none;
-    border-radius: 0.375rem;
-    cursor: pointer;
-    transition: all 0.2s ease;
-}
-
-.basicLayout__settings-button:hover {
-    background-color: var(--color-background-alt);
-}
-
-/* Customize scrollbar appearance */
-.basicLayout__content__lists::-webkit-scrollbar {
-    height: 8px;
-}
-
-.basicLayout__content__lists::-webkit-scrollbar-track {
-    background: var(--color-background-alt);
-    border-radius: 4px;
-}
-
-.basicLayout__content__lists::-webkit-scrollbar-thumb {
-    background: var(--color-background-alt);
-    border-radius: 4px;
-}
-
-.basicLayout__content__lists::-webkit-scrollbar-thumb:hover {
-    background: var(--color-background-alt);
-}
-
-/* Small widget styles */
-.smallWidget {
-    padding: 1rem;
-    height: 100%;
-    overflow-y: auto;
-}
