+++ src/shared/ui/Modal/Modal.css
@@ -26,9 +26,11 @@
 
 .modal__header {
     display: flex;
+    width: 100%;    
     justify-content: space-between;
     align-items: center;
     margin-bottom: 20px;
+    margin-top: auto;
 }
 
 .modal__closeButton {
