+++ src/widgets/InboxList/ui/InboxList.css
@@ -1,5 +1,6 @@
 .inbox-list {
   display: flex;
+  min-width: 300px;
   flex-direction: column;
   gap: 1rem;
   padding: 1rem;
