+++ src/widgets/MainLayout/index.ts
@@ -1,0 +1,1 @@
+export { MainLayout } from './ui/MainLayout';
