+++ src/widgets/MainLayout/ui/MainLayout.css
@@ -1,0 +1,28 @@
+.mainLayout__content__taskList {
+    display: flex;
+    flex-direction: row;
+    gap: 20px;
+    min-width: max-content;
+    padding: 0 1rem;
+}
+
+.baseLayout__content__lists {
+    min-width: max-content;
+
+}
+
+.mainLayout__content__projects,
+.mainLayout__content__goals,
+.mainLayout__content__rewards {
+    width: 400px;
+    min-width: 300px;
+    background: var(--color-background);
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.smallWidget {
+    padding: 1rem;
+    height: 100%;
+    overflow-y: auto;
+}
